requirements.txt
Django==4.0.4
psycopg2-binary==2.9.3
Pillow==9.1.1
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0